Replace Method with BindgenFn
