+++ cs-bindgen-shared/src/bindgen_fn.rs
@@ -1,4 +1,4 @@
-use crate::{FnArg, Primitive, Receiver, ReturnType};
+use crate::{FnArg, Primitive, ReturnType};
 use proc_macro2::Span;
 use serde::*;
 use syn::{spanned::Spanned, *};
@@ -6,6 +6,8 @@ use syn::{spanned::Spanned, *};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct BindgenFn {
     ident: String,
+
+    receiver: Option<Receiver>,
 
     // TODO: Preserve variable names for function arguments or we won't be able to
     // generate code for functions that actually have args.
@@ -15,9 +17,55 @@ pub struct BindgenFn {
 }
 
 impl BindgenFn {
-    pub fn from_item(item: ItemFn) -> syn::Result<Self> {
-        let (ident, _, args, ret) = parse_signature(&item.sig)?;
-        Ok(BindgenFn { ident, args, ret })
+    pub fn from_signature(signature: &Signature) -> syn::Result<Self> {
+        // Generate an error for async functions.
+        if let Some(asyncness) = signature.asyncness {
+            return Err(Error::new(
+                asyncness.span(),
+                "Async functions are not supported with `#[cs_bindgen]`",
+            ));
+        }
+
+        let receiver = signature.receiver().map(Receiver::from_syn).transpose()?;
+
+        // Parse function arguments.
+        let args = signature
+            .inputs
+            .iter()
+            .enumerate()
+            .map(|(index, arg)| match arg {
+                // Reject any functions that take some form of `self`. We'll eventually be able to
+                // support these by marking entire `impl` blocks with `#[cs_bindgen]`, but for now
+                // we only support free functions.
+                syn::FnArg::Receiver(_) => Err(syn::Error::new(
+                    arg.span(),
+                    "Methods are not supported, only free functions",
+                )),
+
+                syn::FnArg::Typed(pat) => {
+                    // If the argument isn't declared with a normal identifier, we construct one so
+                    // that we have a valid identifier to use in the generated functions.
+                    let ident = match &*pat.pat {
+                        Pat::Ident(pat_ident) => pat_ident.ident.to_string(),
+                        _ => format!("__arg{}", index),
+                    };
+
+                    let ty = Primitive::from_type(&pat.ty).ok_or(syn::Error::new(
+                        pat.ty.span(),
+                        "Unknown argument type, only primitives are supported",
+                    ))?;
+
+                    Ok(FnArg::new(ident, ty))
+                }
+            })
+            .collect::<syn::Result<_>>()?;
+
+        Ok(Self {
+            ident: signature.ident.to_string(),
+            receiver,
+            args,
+            ret: ReturnType::from_syn(&signature.output)?,
+        })
     }
 
     /// Returns the raw string representation of the function's name.
@@ -44,55 +92,32 @@ impl BindgenFn {
     }
 }
 
-pub fn parse_signature(
-    signature: &Signature,
-) -> syn::Result<(String, Option<Receiver>, Vec<FnArg>, ReturnType)> {
-    // Generate an error for async functions.
-    if let Some(asyncness) = signature.asyncness {
-        return Err(Error::new(
-            asyncness.span(),
-            "Async functions are not supported with `#[cs_bindgen]`",
-        ));
-    }
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub enum Receiver {
+    Ref,
+    RefMut,
+    Value,
+}
 
-    let receiver = signature.receiver().map(Receiver::from_syn).transpose()?;
+impl Receiver {
+    pub fn from_syn(arg: &syn::FnArg) -> syn::Result<Self> {
+        match arg {
+            syn::FnArg::Receiver(recv) => Ok({
+                if recv.reference.is_some() {
+                    if recv.mutability.is_some() {
+                        Receiver::RefMut
+                    } else {
+                        Receiver::Ref
+                    }
+                } else {
+                    Receiver::Value
+                }
+            }),
 
-    // Parse function arguments.
-    let args = signature
-        .inputs
-        .iter()
-        .enumerate()
-        .map(|(index, arg)| match arg {
-            // Reject any functions that take some form of `self`. We'll eventually be able to
-            // support these by marking entire `impl` blocks with `#[cs_bindgen]`, but for now
-            // we only support free functions.
-            syn::FnArg::Receiver(_) => Err(syn::Error::new(
+            syn::FnArg::Typed(_) => Err(Error::new(
                 arg.span(),
-                "Methods are not supported, only free functions",
+                "Invalid receiver for `#[cs_bindgen]` method",
             )),
-
-            syn::FnArg::Typed(pat) => {
-                // If the argument isn't declared with a normal identifier, we construct one so
-                // that we have a valid identifier to use in the generated functions.
-                let ident = match &*pat.pat {
-                    Pat::Ident(pat_ident) => pat_ident.ident.to_string(),
-                    _ => format!("__arg{}", index),
-                };
-
-                let ty = Primitive::from_type(&pat.ty).ok_or(syn::Error::new(
-                    pat.ty.span(),
-                    "Unknown argument type, only primitives are supported",
-                ))?;
-
-                Ok(FnArg::new(ident, ty))
-            }
-        })
-        .collect::<syn::Result<_>>()?;
-
-    Ok((
-        signature.ident.to_string(),
-        receiver,
-        args,
-        ReturnType::from_syn(&signature.output)?,
-    ))
+        }
+    }
 }
