+++ cs-bindgen-shared/src/bindgen_impl.rs
@@ -1,11 +1,11 @@
-use crate::{parse_signature, FnArg, ReturnType};
+use crate::BindgenFn;
 use serde::*;
-use syn::{spanned::Spanned, Error, ImplItem, ImplItemMethod, Type};
+use syn::{spanned::Spanned, Error, ImplItem, Type};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct BindgenImpl {
     pub ty_ident: String,
-    pub methods: Vec<Method>,
+    pub methods: Vec<BindgenFn>,
 }
 
 impl BindgenImpl {
@@ -41,73 +41,14 @@ impl BindgenImpl {
 
         let methods = item
             .items
-            .into_iter()
+            .iter()
             .filter_map(|item| match item {
-                ImplItem::Method(item) => Some(item),
+                ImplItem::Method(item) => Some(&item.sig),
                 _ => None,
             })
-            .map(Method::from_item)
+            .map(BindgenFn::from_signature)
             .collect::<syn::Result<_>>()?;
 
         Ok(Self { ty_ident, methods })
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Method {
-    ident: String,
-
-    receiver: Receiver,
-
-    // TODO: Preserve variable names for function arguments or we won't be able to
-    // generate code for functions that actually have args.
-    pub args: Vec<FnArg>,
-
-    pub ret: ReturnType,
-}
-
-impl Method {
-    pub fn from_item(item: ImplItemMethod) -> syn::Result<Self> {
-        let (ident, receiver, args, ret) = parse_signature(&item.sig)?;
-
-        Ok(Method {
-            ident,
-            receiver: receiver.ok_or(Error::new(
-                item.sig.span(),
-                "Invalid receiver for `#[cs_bindgen]` method",
-            ))?,
-            args,
-            ret,
-        })
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum Receiver {
-    Ref,
-    RefMut,
-    Value,
-}
-
-impl Receiver {
-    pub fn from_syn(arg: &syn::FnArg) -> syn::Result<Self> {
-        match arg {
-            syn::FnArg::Receiver(recv) => Ok({
-                if recv.reference.is_some() {
-                    if recv.mutability.is_some() {
-                        Receiver::RefMut
-                    } else {
-                        Receiver::Ref
-                    }
-                } else {
-                    Receiver::Value
-                }
-            }),
-
-            syn::FnArg::Typed(_) => Err(Error::new(
-                arg.span(),
-                "Invalid receiver for `#[cs_bindgen]` method",
-            )),
-        }
     }
 }
