+++ cs-bindgen-shared/src/lib.rs
@@ -8,8 +8,8 @@ mod ret;
 
 pub use crate::{
     arg::FnArg,
-    bindgen_fn::{parse_signature, BindgenFn},
-    bindgen_impl::{BindgenImpl, Method, Receiver},
+    bindgen_fn::{BindgenFn, Receiver},
+    bindgen_impl::BindgenImpl,
     bindgen_struct::BindgenStruct,
     item::BindgenItem,
     primitive::Primitive,
